+++ Realm/RLMRealm.mm
@@ -612,9 +612,9 @@ void RLMRealmTranslateException(NSError **error) {
 }
 
 - (void)addOrUpdateObject:(RLMObject *)object {
-    // verify primary key
-    if (!object.objectSchema.primaryKeyProperty) {
-        @throw RLMException(@"'%@' does not have a primary key and can not be updated", object.objectSchema.className);
+    // verify object ID
+    if (!object.objectSchema.objectIDProperty) {
+        @throw RLMException(@"'%@' does not have an object ID and can not be updated", object.objectSchema.className);
     }
 
     RLMAddObjectToRealm(object, self, true);
@@ -675,6 +675,10 @@ void RLMRealmTranslateException(NSError **error) {
 
 - (RLMObject *)objectWithClassName:(NSString *)className forPrimaryKey:(id)primaryKey {
     return RLMGetObject(self, className, primaryKey);
+}
+
+- (RLMObject *)objectWithClassName:(NSString *)className withID:(id)objectID {
+    return RLMGetObject(self, className, objectID);
 }
 
 + (uint64_t)schemaVersionAtPath:(NSString *)realmPath error:(NSError **)error {