+++ Realm/RLMRealm_Dynamic.h
@@ -104,7 +104,27 @@
  
  @see       RLMObject objectForPrimaryKey:
  */
-- (RLMObject *)objectWithClassName:(NSString *)className forPrimaryKey:(id)primaryKey;
+- (RLMObject *)objectWithClassName:(NSString *)className forPrimaryKey:(id)primaryKey DEPRECATED_MSG_ATTRIBUTE("use -[RLMRealm objectWithClassName:withID:]");
+
+/**
+ This method is useful only in specialized circumstances, for example, when building components
+ that integrate with Realm. If you are simply building an app on Realm, it is
+ recommended to use the class methods on `RLMObject`.
+ 
+ Get an object of a given class name with an object ID
+ 
+ The preferred way to get an object of a single class is to use the class methods on RLMObject.
+ 
+ @warning This method is useful only in specialized circumstances.
+ 
+ @param className   The class name for the object you are looking for
+ @param objectID  The object ID value for the object you are looking for
+ 
+ @return    An object or nil if an object with the given object ID does not exist.
+ 
+ @see       RLMObject objectWithID:
+ */
+- (RLMObject *)objectWithClassName:(NSString *)className withID:(id)objectID;
 
 /**
  This method is useful only in specialized circumstances, for example, when building components