+++ Realm/Tests/RLMTestObjects.m
@@ -158,7 +158,7 @@
 #pragma mark PrimaryStringObject
 
 @implementation PrimaryStringObject
-+ (NSString *)primaryKey {
++ (NSString *)objectID {
     return @"stringCol";
 }
 @end
@@ -215,5 +215,6 @@
 + (NSArray *)ignoredProperties { return nil; }
 + (NSArray *)indexedProperties { return nil; }
 + (NSString *)primaryKey { return nil; }
++ (NSString *)objectID { return nil; }
 + (NSArray *)requiredProperties { return nil; }
 @end