+++ Realm/Tests/Swift2.0/SwiftTestObjects.swift
@@ -123,7 +123,7 @@ class SwiftPrimaryStringObject: RLMObject {
     dynamic var stringCol = ""
     dynamic var intCol = 0
 
-    override class func primaryKey() -> String {
+    override class func objectID() -> String {
         return "stringCol"
     }
 }