+++ RealmSwift-swift2.0/Tests/SwiftTestObjects.swift
@@ -200,7 +200,7 @@ class SwiftLinkToPrimaryStringObject: Object {
     dynamic var object: SwiftPrimaryStringObject?
     let objects = List<SwiftPrimaryStringObject>()
 
-    override class func primaryKey() -> String? {
+    override class func objectID() -> String? {
         return "pk"
     }
 }
@@ -230,7 +230,7 @@ class SwiftPrimaryStringObject: Object {
     dynamic var stringCol = ""
     dynamic var intCol = 0
 
-    override class func primaryKey() -> String? {
+    override class func objectID() -> String? {
         return "stringCol"
     }
 }